backend/requirements.txt
asgiref==3.8.1
astroid==3.2.4
autopep8==2.3.1
dill==0.3.8
Django==4.2.15
django-cors-headers==4.4.0
django-enumfields==2.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
isort==5.13.2
jedi==0.19.1
mccabe==0.7.0
parso==0.8.4
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycodestyle==2.12.1
PyJWT==2.9.0
pylint==3.2.6
python-dotenv==1.0.1
sqlparse==0.5.1
tomli==2.0.1
tomlkit==0.13.0
typing-extensions==4.12.2
whitenoise==6.7.0
